feat: compliance questions and popup

+++ src/Utils/dateTime.ts
@@ -9,7 +9,7 @@ export function getStandatedDateWithTime(dateString : string) {
 export function getStandatedDate(dateString : string) {
 
     const date = new Date(dateString);
-    const options : any= { year: 'numeric', month: 'long', day: 'numeric'};
+    const options : any= { year: 'numeric', month: 'short', day: 'numeric'};
     const formattedDate = date.toLocaleDateString('en-US', options );
     return formattedDate
 }
